requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
typing-extensions
pytest>=7.4.0
pytest-cov>=4.0.0